rest_api/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.1
django-extensions==4.1
Werkzeug==3.1.3
pyOpenSSL==25.1.0
requests>=2.31.0
aio_pika
httpx
python-dateutil
Pillow